--- pages/api/works.ts
@@ -1,19 +1,0 @@
-import { NextApiRequest, NextApiResponse } from "next";
-import { prisma } from "../../lib/prisma";
-
-export default async function handler(
-    req: NextApiRequest,
-    res: NextApiResponse
-) {
-    if(req.method == 'GET') {
-        try {
-            const works = await prisma.work.findMany();
-            res.status(200).json(works);
-        } catch (e) {
-            res.status(500).json({ error: 'Internal Server Error' });
-        }
-    } else {
-        res.setHeader('Allow', ['GET']);
-        res.status(405).end(`Method ${req.method} Not Allowed`);
-    }
-}
