+++ src/pages/api/contactOthers.ts
@@ -1,5 +1,5 @@
 import { NextApiRequest, NextApiResponse } from "next";
-import { prisma } from "../../lib/prisma";
+import { prisma } from "../../../lib/prisma";
 
 export default async function handler(
   req: NextApiRequest,
@@ -7,9 +7,8 @@ export default async function handler(
 ) {
   if (req.method == "GET") {
     try {
-      const contactEmailAndPhones =
-        await prisma.contactEmailAndPhone.findMany();
-      res.status(200).json(contactEmailAndPhones);
+      const contactOthers = await prisma.contactOthers.findMany();
+      res.status(200).json(contactOthers);
     } catch (e) {
       res.status(500).json({ error: "Internak Server Error" });
     }
