+++ src/pages/api/skillLanguageProficiencies.ts
@@ -1,5 +1,5 @@
 import { NextApiRequest, NextApiResponse } from "next";
-import { prisma } from "../../lib/prisma";
+import { prisma } from "../../../lib/prisma";
 
 export default async function handler(
   req: NextApiRequest,
@@ -7,9 +7,9 @@ export default async function handler(
 ) {
   if (req.method == "GET") {
     try {
-      const skillFrameWorkProficiencies =
-        await prisma.skillFrameworkProficiency.findMany();
-      res.status(200).json(skillFrameWorkProficiencies);
+      const skillLanguageProficiencies =
+        await prisma.skillLanguageProficiency.findMany();
+      res.status(200).json(skillLanguageProficiencies);
     } catch (e) {
       res.status(500).json({ error: "Internak Server Error" });
     }
