+++ src/pages/contests/index.tsx
@@ -1,8 +1,9 @@
-import Title from "@/components/title";
 import ContestCardGrid from "./contests_card_grid";
 import { useTheme } from "@mui/material/styles";
 import ContestsText from "@/components/contests/contests_text";
 import ContestsDetailText from "@/components/contests/contests_detail_text";
+import Header from "../header";
+import Footer from "../footer";
 
 const Contests = () => {
   const theme = useTheme();
@@ -12,6 +13,7 @@ const Contests = () => {
         backgroundColor: theme.palette.mode === "dark" ? "#18181B" : "white",
       }}
     >
+      <Header />
       <div style={{ width: "60%", padding: "40px" }}>
         <ContestsText
           text={"Mobile developer working primarily with Flutter"}
@@ -23,6 +25,7 @@ const Contests = () => {
         />
       </div>
       <ContestCardGrid />
+      <Footer />
     </div>
   );
 };
